Move multiplatform config to compose handler (including plugin)
